tools/agentRag/requirements.txt
# Core dependencies
flask==3.0.0
fastapi==0.104.1
uvicorn==0.24.0
aiohttp==3.9.0

# API clients
google-generativeai==0.3.0
anthropic==0.18.0
openai==1.10.0  # For Perplexity API compatibility

# Text processing
tiktoken==0.5.1
python-slugify==8.0.1

# Database
sqlite3-api==0.1.0  # SQLite utilities

# Optional: Embeddings
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
logging-config==1.0.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
pylint==3.0.2

# Optional: Web UI
streamlit==1.28.0
gradio==4.5.0

# Optional: MCP Runtime (if using Node.js version)
# Install separately: npm install -g @modelcontextprotocol/runtime